+++ Example/AppDelegate.swift
@@ -22,12 +22,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         loggerPlugin.logSuccess = true
         tron = TRON(baseURL: "https://api.github.com", plugins: [loggerPlugin])
         tron.headerBuilder = HeaderBuilder(defaultHeaders: [:])
-        let request : APIRequest<String,Int> = tron.request("zen")
+        let request : APIRequest<String,Int> = tron.swiftyJSON.request("zen")
         let token = request.perform(withSuccess: { zen in
             print(zen)
             }, failure: { error in
         })
-        debugPrint(token)
+        debugPrint(token ?? "")
         // Override point for customization after application launch.
         return true
     }